Added the possibility of sending parameters to algorithms.
Now, if two objects use the same segmentation algorithm, but with different parameters, two instances of this algorithm will be made and executed, as
if the objects used different segmentation methods. Also, the base classes for the algorithms have an "init" method, that will be used to hold anything
that needs to be done before the main loop.